+++ src/android/PESDKPlugin.kt
@@ -119,7 +119,7 @@ class PESDKPlugin : CordovaPlugin() {
         MainThreadRunnable {
             EditorBuilder(currentActivity)
               .setSettingsList(settingsList)
-              .startActivityForResult(currentActivity, EDITOR_RESULT_ID)
+              .startActivityForResult(currentActivity, EDITOR_RESULT_ID, arrayOfNulls(0))
         }()
     }